requirements.txt
tensorflow==1.15.2
pillow
lxml
jupyter
matplotlib
opencv-python
cython
pandas
imageio